refactor: FoodList 컴포넌트 자식 컴포넌트 모듈화 완료
