+++ frontend/src/pages/FoodList/FoodTrucks/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { FoodListContainer, FoodListWrapper } from './styles';
-import { FoodTruck } from '../../components';
+import { FoodTruck } from '../../../components';
 
 function FoodTrucks() {
   const data = [1, 2, 3, 4];
